Move I2C accessors to separate type
